move interface from class to module
